+++ filament/backend/src/opengl/OpenGLDriver.cpp
@@ -231,6 +231,8 @@ void OpenGLDriver::terminate() {
     // wait for the GPU to finish executing all commands
     glFinish();
 
+    mShaderCompilerService.terminate();
+
     // and make sure to execute all the GpuCommandCompleteOps callbacks
     executeGpuCommandsCompleteOps();
 
@@ -248,8 +250,6 @@ void OpenGLDriver::terminate() {
 #endif
 
     delete mTimerQueryImpl;
-
-    mShaderCompilerService.terminate();
 
     mPlatform.terminate();
 }
@@ -2607,10 +2607,10 @@ SyncStatus OpenGLDriver::getSyncStatus(Handle<HwSync> sh) {
     }
 }
 
-void OpenGLDriver::compilePrograms(CallbackHandler* handler,
-        CallbackHandler::Callback callback, void* user) {
+void OpenGLDriver::compilePrograms(CompilerPriorityQueue priority,
+        CallbackHandler* handler, CallbackHandler::Callback callback, void* user) {
     if (callback) {
-        getShaderCompilerService().notifyWhenAllProgramsAreReady(handler, callback, user);
+        getShaderCompilerService().notifyWhenAllProgramsAreReady(priority, handler, callback, user);
     }
 }