+++ filament/src/details/Engine.h
@@ -293,6 +293,26 @@ public:
     bool destroy(const FView* p);
     bool destroy(const FInstanceBuffer* p);
 
+    bool isValid(const FBufferObject* p);
+    bool isValid(const FVertexBuffer* p);
+    bool isValid(const FFence* p);
+    bool isValid(const FIndexBuffer* p);
+    bool isValid(const FSkinningBuffer* p);
+    bool isValid(const FMorphTargetBuffer* p);
+    bool isValid(const FIndirectLight* p);
+    bool isValid(const FMaterial* p);
+    bool isValid(const FMaterialInstance* p);
+    bool isValid(const FRenderer* p);
+    bool isValid(const FScene* p);
+    bool isValid(const FSkybox* p);
+    bool isValid(const FColorGrading* p);
+    bool isValid(const FSwapChain* p);
+    bool isValid(const FStream* p);
+    bool isValid(const FTexture* p);
+    bool isValid(const FRenderTarget* p);
+    bool isValid(const FView* p);
+    bool isValid(const FInstanceBuffer* p);
+
     void destroy(utils::Entity e);
 
     void flushAndWait();
@@ -391,6 +411,9 @@ private:
     void flushCommandBuffer(backend::CommandBufferQueue& commandBufferQueue);
 
     backend::Driver& getDriver() const noexcept { return *mDriver; }
+
+    template<typename T>
+    bool isValid(const T* ptr, ResourceList<T>& list);
 
     template<typename T>
     bool terminateAndDestroy(const T* p, ResourceList<T>& list);