+++ libs/filamat/src/GLSLPostProcessor.cpp
@@ -335,7 +335,7 @@ bool GLSLPostProcessor::process(const std::string& inputShader, Config const& co
 
     // This allows shaders to query if they will be run through glslang.
     // OpenGL shaders without optimization, for example, won't have this define.
-    tShader.setPreamble("#define FILAMENT_GLSLANG 1\n");
+    tShader.setPreamble("#define FILAMENT_GLSLANG\n");
 
     internalConfig.langVersion = GLSLTools::getGlslDefaultVersion(config.shaderModel);
     GLSLTools::prepareShaderParser(config.targetApi, config.targetLanguage, tShader,
@@ -616,9 +616,11 @@ void GLSLPostProcessor::fixupClipDistance(
     if (!config.usesClipDistance) {
         return;
     }
+    // This should match the version of SPIR-V used in GLSLTools::prepareShaderParser.
     SpirvTools tools(SPV_ENV_UNIVERSAL_1_3);
     std::string disassembly;
-    tools.Disassemble(spirv, &disassembly);
+    const bool result = tools.Disassemble(spirv, &disassembly);
+    assert_invariant(result);
     if (filamat::fixupClipDistance(disassembly)) {
         spirv.clear();
         tools.Assemble(disassembly, &spirv);